+++ src/main/kotlin/com/mituuz/fuzzier/FuzzyGrep.kt
@@ -28,8 +28,8 @@ package com.mituuz.fuzzier
 
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.execution.process.OSProcessHandler
-import com.intellij.execution.process.ProcessAdapter
 import com.intellij.execution.process.ProcessEvent
+import com.intellij.execution.process.ProcessListener
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationType
 import com.intellij.notification.Notifications
@@ -40,9 +40,6 @@ import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.LogicalPosition
 import com.intellij.openapi.editor.ScrollType
 import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.progress.ProgressManager
-import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.JBPopup
 import com.intellij.openapi.ui.popup.JBPopupListener
@@ -59,42 +56,46 @@ import java.awt.event.ActionEvent
 import java.awt.event.KeyEvent
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
-import java.io.IOException
-import java.util.concurrent.TimeUnit
-import java.util.concurrent.locks.ReentrantLock
 import javax.swing.AbstractAction
 import javax.swing.DefaultListModel
 import javax.swing.JComponent
 import javax.swing.KeyStroke
-import kotlin.coroutines.cancellation.CancellationException
 
 open class FuzzyGrep() : FuzzyAction() {
     companion object {
         const val FUZZIER_NOTIFICATION_GROUP: String = "Fuzzier Notification Group"
+
+        /**
+         * Limit command output size, this is only used to check installations
+         */
+        const val MAX_OUTPUT_SIZE = 10000
+        const val MAX_NUMBER_OR_RESULTS = 1000
     }
 
     override var popupTitle: String = "Fuzzy Grep"
     override var dimensionKey = "FuzzyGrepPopup"
-    private var lock = ReentrantLock()
     var useRg = true
     val isWindows = System.getProperty("os.name").lowercase().contains("win")
-    var currentJob: Job? = null
+    private var currentLaunchJob: Job? = null
+    private var currentUpdateListContentJob: Job? = null
+    private var actionScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
 
     override fun runAction(
         project: Project,
         actionEvent: AnActionEvent
     ) {
-        currentJob?.cancel()
+        currentLaunchJob?.cancel()
         setCustomHandlers()
 
         val projectBasePath = project.basePath.toString()
-        currentJob = CoroutineScope(Dispatchers.EDT).launch {
-            val rgCommand = checkInstallation("rg", projectBasePath)
-            if (rgCommand != null) {
+        currentLaunchJob = actionScope.launch(Dispatchers.EDT) {
+            val currentJob = currentLaunchJob
+
+            if (!isInstalled("rg", projectBasePath)) {
                 showNotification(
                     "No `rg` command found",
                     """
-                    No ripgrep found with command: $rgCommand<br>
+                    No ripgrep found<br>
                     Fallback to `grep` or `findstr`<br>
                     This notification can be disabled
                 """.trimIndent(),
@@ -103,22 +104,20 @@ open class FuzzyGrep() : FuzzyAction() {
                 )
 
                 if (isWindows) {
-                    val findstrCommand = checkInstallation("findstr", projectBasePath)
-                    if (findstrCommand != null) {
+                    if (!isInstalled("findstr", projectBasePath)) {
                         showNotification(
                             "No `findstr` command found",
-                            "No findstr found with command: $findstrCommand",
+                            "Fuzzy Grep failed: no `findstr` found",
                             project
                         )
                         return@launch
                     }
                     popupTitle = "Fuzzy Grep (findstr)"
                 } else {
-                    val grepCommand = checkInstallation("grep", projectBasePath)
-                    if (grepCommand != null) {
+                    if (!isInstalled("grep", projectBasePath)) {
                         showNotification(
                             "No `grep` command found",
-                            "No grep found with command: $grepCommand",
+                            "Fuzzy Grep failed: no `grep` found",
                             project
                         )
                         return@launch
@@ -130,9 +129,10 @@ open class FuzzyGrep() : FuzzyAction() {
                 popupTitle = "Fuzzy Grep (ripgrep)"
                 useRg = true
             }
-        }
 
-        ApplicationManager.getApplication().invokeLater {
+            if (currentJob?.isCancelled == true) return@launch
+
+            yield()
             defaultDoc = EditorFactory.getInstance().createDocument("")
             component = FuzzyFinderComponent(project)
             createListeners(project)
@@ -169,7 +169,8 @@ open class FuzzyGrep() : FuzzyAction() {
                     (component as FuzzyFinderComponent).splitPane.dividerLocation
                 resetOriginalHandlers()
                 super.onClosed(event)
-                currentTask?.cancel(true)
+                currentLaunchJob?.cancel()
+                currentUpdateListContentJob?.cancel()
             }
         })
 
@@ -178,21 +179,18 @@ open class FuzzyGrep() : FuzzyAction() {
 
     /**
      * OS-specific to see if a specific executable is found
-     * @return the used command if no installation detected, otherwise null
+     * @return true if the command was found (possibly not a general check/solution)
      */
-    private suspend fun checkInstallation(executable: String, projectBasePath: String): String? {
+    private suspend fun isInstalled(executable: String, projectBasePath: String): Boolean {
         val command = if (isWindows) {
             listOf("where", executable)
         } else {
             listOf("which", executable)
         }
 
-        val result = withContext(Dispatchers.IO) { runCommand(command, projectBasePath) }
-        if (result.isNullOrBlank() || result.contains("Could not find files")) {
-            return command.joinToString(" ")
-        }
+        val result = runCommand(command, projectBasePath)
 
-        return null
+        return !(result.isNullOrBlank() || result.contains("Could not find files"))
     }
 
     override fun updateListContents(project: Project, searchString: String) {
@@ -201,80 +199,111 @@ open class FuzzyGrep() : FuzzyAction() {
             return
         }
 
-        currentTask?.takeIf { !it.isDone }?.cancel(true)
-        if (lock.tryLock(500, TimeUnit.MILLISECONDS)) {
+        currentUpdateListContentJob?.cancel()
+        currentUpdateListContentJob = actionScope.launch(Dispatchers.EDT) {
+            val currentJob = currentUpdateListContentJob
 
-            try {
-                currentTask = ApplicationManager.getApplication().executeOnPooledThread {
-                    try {
-                        val task = currentTask
+            if (currentJob?.isCancelled == true) return@launch
 
-                        if (task?.isCancelled == true) return@executeOnPooledThread
+            component.fileList.setPaintBusy(true)
+            val listModel = DefaultListModel<FuzzyContainer>()
 
-                        component.fileList.setPaintBusy(true)
-                        val listModel = DefaultListModel<FuzzyContainer>()
+            if (currentJob?.isCancelled == true) return@launch
 
-                        if (task?.isCancelled == true) return@executeOnPooledThread
-
-                        findInFiles(searchString, listModel, project.basePath.toString())
-
-                        if (task?.isCancelled == true) return@executeOnPooledThread
-
-                        ApplicationManager.getApplication().invokeLater {
-                            synchronized(component.fileList.model) {
-                                component.fileList.model = listModel
-                                component.fileList.cellRenderer = getCellRenderer()
-                                if (!listModel.isEmpty) {
-                                    component.fileList.selectedIndex = 0
-                                }
-                                component.fileList.setPaintBusy(false)
-                            }
-                        }
-                    } catch (_: InterruptedException) {
-                        return@executeOnPooledThread
-                    } catch (_: CancellationException) {
-                        return@executeOnPooledThread
-                    }
-                }
-                try {
-                    currentTask?.get()
-                } catch (_: InterruptedException) {
-                } catch (_: CancellationException) {
-                }
-            } finally {
-                lock.unlock()
+            val results = withContext(Dispatchers.IO) {
+                findInFiles(searchString, listModel, project.basePath.toString())
+                listModel
             }
+
+            if (currentJob?.isCancelled == true) return@launch
+
+            component.fileList.model = results
+            component.fileList.cellRenderer = getCellRenderer()
+            if (!results.isEmpty) {
+                component.fileList.selectedIndex = 0
+            }
+            component.fileList.setPaintBusy(false)
         }
     }
 
-    protected open fun runCommand(commands: List<String>, projectBasePath: String): String? {
+    /**
+     * Run the command and collect the output to a string variable with a limited size
+     * @see MAX_OUTPUT_SIZE
+     */
+    protected open suspend fun runCommand(commands: List<String>, projectBasePath: String): String? {
         return try {
             val commandLine = GeneralCommandLine(commands)
                 .withWorkDirectory(projectBasePath)
                 .withRedirectErrorStream(true)
             val output = StringBuilder()
             val processHandler = OSProcessHandler(commandLine)
-            processHandler.addProcessListener(object : ProcessAdapter() {
+
+            processHandler.addProcessListener(object : ProcessListener {
                 override fun onTextAvailable(event: ProcessEvent, outputType: Key<*>) {
-                    output.appendLine(event.text.replace("\n", ""))
+                    if (output.length < MAX_OUTPUT_SIZE) {
+                        output.appendLine(event.text.replace("\n", ""))
+                    }
                 }
             })
-            processHandler.startNotify()
-            processHandler.waitFor(2000)
+
+            withContext(Dispatchers.IO) {
+                processHandler.startNotify()
+                processHandler.waitFor(2000)
+            }
             output.toString()
-        } catch (_: IOException) {
-            null
         } catch (_: InterruptedException) {
-            null
+            throw InterruptedException()
         }
     }
 
-    private fun findInFiles(
+    /**
+     * Run the command and stream a limited number of results to the list model
+     * @see MAX_NUMBER_OR_RESULTS
+     */
+    protected open suspend fun runCommand(
+        commands: List<String>,
+        listModel: DefaultListModel<FuzzyContainer>,
+        projectBasePath: String
+    ) {
+        try {
+            val commandLine = GeneralCommandLine(commands)
+                .withWorkDirectory(projectBasePath)
+                .withRedirectErrorStream(true)
+
+            val processHandler = OSProcessHandler(commandLine)
+            var count = 0
+
+            processHandler.addProcessListener(object : ProcessListener {
+                override fun onTextAvailable(event: ProcessEvent, outputType: Key<*>) {
+                    if (count >= MAX_NUMBER_OR_RESULTS) return
+
+                    event.text.lines().forEach { line ->
+                        if (count >= MAX_NUMBER_OR_RESULTS) return@forEach
+                        if (line.isNotBlank()) {
+                            val rowContainer = RowContainer.rowContainerFromString(line, projectBasePath, useRg)
+                            if (rowContainer != null) {
+                                listModel.addElement(rowContainer)
+                                count++
+                            }
+                        }
+                    }
+                }
+            })
+
+            withContext(Dispatchers.IO) {
+                processHandler.startNotify()
+                processHandler.waitFor(2000)
+            }
+        } catch (_: InterruptedException) {
+            throw InterruptedException()
+        }
+    }
+
+    private suspend fun findInFiles(
         searchString: String, listModel: DefaultListModel<FuzzyContainer>,
         projectBasePath: String
     ) {
-
-        val res = if (useRg) {
+        if (useRg) {
             runCommand(
                 listOf(
                     "rg",
@@ -285,26 +314,14 @@ open class FuzzyGrep() : FuzzyAction() {
                     "--column",
                     searchString,
                     "."
-                ), projectBasePath
+                ), listModel, projectBasePath
             )
         } else {
             if (isWindows) {
-                runCommand(listOf("findstr", "/p", "/s", "/n", searchString, "*"), projectBasePath)
+                runCommand(listOf("findstr", "/p", "/s", "/n", searchString, "*"), listModel, projectBasePath)
             } else {
-                runCommand(listOf("grep", "--color=none", "-r", "-n", searchString, "."), projectBasePath)
+                runCommand(listOf("grep", "--color=none", "-r", "-n", searchString, "."), listModel, projectBasePath)
             }
-        }
-
-        if (res != null) {
-            res.lines()
-                .forEach { line ->
-                    val rowContainer = RowContainer.rowContainerFromString(line, projectBasePath, useRg)
-                    if (rowContainer != null) {
-                        listModel.addElement(rowContainer)
-                    }
-                }
-        } else {
-            return
         }
     }
 
@@ -313,7 +330,7 @@ open class FuzzyGrep() : FuzzyAction() {
         component.fileList.addListSelectionListener { event ->
             if (!event.valueIsAdjusting) {
                 if (component.fileList.isEmpty) {
-                    ApplicationManager.getApplication().invokeLater {
+                    actionScope.launch(Dispatchers.EDT) {
                         defaultDoc?.let { (component as FuzzyFinderComponent).previewPane.updateFile(it) }
                     }
                     return@addListSelectionListener
@@ -321,17 +338,18 @@ open class FuzzyGrep() : FuzzyAction() {
                 val selectedValue = component.fileList.selectedValue
                 val fileUrl = "file://${selectedValue?.getFileUri()}"
 
-                ProgressManager.getInstance().run(object : Task.Backgroundable(null, "Loading file", false) {
-                    override fun run(indicator: ProgressIndicator) {
-                        val file = VirtualFileManager.getInstance().findFileByUrl(fileUrl)
-                        file?.let {
-                            (component as FuzzyFinderComponent).previewPane.updateFile(
-                                file,
-                                (selectedValue as RowContainer).rowNumber
-                            )
-                        }
+                actionScope.launch(Dispatchers.Default) {
+                    val file = withContext(Dispatchers.IO) {
+                        VirtualFileManager.getInstance().findFileByUrl(fileUrl)
                     }
-                })
+
+                    file?.let {
+                        (component as FuzzyFinderComponent).previewPane.coUpdateFile(
+                            file,
+                            (selectedValue as RowContainer).rowNumber
+                        )
+                    }
+                }
             }
         }