+++ src/main/kotlin/com/mituuz/fuzzier/components/PreviewEditor.kt
@@ -26,6 +26,8 @@
 package com.mituuz.fuzzier.components
 
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.application.EDT
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Document
@@ -34,6 +36,7 @@ import com.intellij.openapi.editor.LogicalPosition
 import com.intellij.openapi.editor.ScrollType
 import com.intellij.openapi.editor.colors.EditorColorsManager
 import com.intellij.openapi.editor.ex.EditorEx
+import com.intellij.openapi.editor.highlighter.EditorHighlighter
 import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.fileTypes.FileType
@@ -45,13 +48,17 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.EditorTextField
 import com.intellij.ui.JBColor
 import com.mituuz.fuzzier.settings.FuzzierGlobalSettingsService
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import kotlin.math.max
 import kotlin.math.min
 
 class PreviewEditor(project: Project?) : EditorTextField(
     project,
     getDefaultFileType()
 ) {
-    private val toBeContinued: String = "\n\n\n--- End of Fuzzier Preview ---\n--- Open file to see full content ---\n\n\n"
+    private val toBeContinued: String =
+        "\n\n\n--- End of Fuzzier Preview ---\n--- Open file to see full content ---\n\n\n"
     private val fileCutOff: Int = 70000
     private val globalState = service<FuzzierGlobalSettingsService>().state
 
@@ -86,7 +93,52 @@ class PreviewEditor(project: Project?) : EditorTextField(
         this.fileType = PlainTextFileType.INSTANCE
     }
 
-    fun updateFile(virtualFile: VirtualFile?, rowNumber: Int? = null) {
+    suspend fun coUpdateFile(virtualFile: VirtualFile, rowNumber: Int) {
+        val sourceDocument = withContext(Dispatchers.IO) {
+            readAction {
+                virtualFile.let { FileDocumentManager.getInstance().getDocument(virtualFile) }
+            }
+        }
+
+        val fileType = virtualFile.let { FileTypeManager.getInstance().getFileTypeByFile(virtualFile) }
+
+        withContext(Dispatchers.EDT) {
+            if (sourceDocument != null) {
+                val previewLines = 100
+                val linesAround = previewLines / 2
+
+                val totalLines = sourceDocument.lineCount
+                val startLine = max(0, rowNumber - linesAround)
+                val endLine = min(totalLines - 1, rowNumber + linesAround)
+                val actualPosition = rowNumber - startLine
+
+                val startIndex = sourceDocument.getLineStartOffset(startLine)
+                val endIndex = sourceDocument.getLineEndOffset(endLine)
+
+                val previewText = withContext(Dispatchers.IO) {
+                    readAction {
+                        sourceDocument.getText(TextRange(startIndex, endIndex))
+                    }
+                }
+
+                this@PreviewEditor.document = EditorFactory.getInstance().createDocument(previewText)
+                this@PreviewEditor.fileType = fileType
+
+                editor?.scrollingModel?.run {
+                    scrollTo(LogicalPosition(actualPosition, 0), ScrollType.CENTER)
+                    val markupModel = editor?.markupModel
+                    val textAttributes = TextAttributes()
+                    textAttributes.backgroundColor = JBColor.LIGHT_GRAY
+                    markupModel?.addLineHighlighter(actualPosition, actualPosition + 1, textAttributes)
+                }
+            } else {
+                this@PreviewEditor.document = EditorFactory.getInstance().createDocument("Cannot preview file")
+                this@PreviewEditor.fileType = fileType
+            }
+        }
+    }
+
+    fun updateFile(virtualFile: VirtualFile?) {
         ApplicationManager.getApplication().executeOnPooledThread {
             val sourceDocument = ApplicationManager.getApplication().runReadAction<Document?> {
                 virtualFile?.let { FileDocumentManager.getInstance().getDocument(virtualFile) }
@@ -109,17 +161,7 @@ class PreviewEditor(project: Project?) : EditorTextField(
                         }
                         ApplicationManager.getApplication().invokeLater {
                             editor?.scrollingModel?.run {
-                                if (rowNumber != null) {
-                                    scrollTo(LogicalPosition(rowNumber, 0), ScrollType.CENTER)
-                                    val markupModel = editor?.markupModel
-                                    val textAttributes = TextAttributes()
-                                    textAttributes.backgroundColor = JBColor.LIGHT_GRAY
-                                    if (rowNumber in 0 until document.lineCount) {
-                                        markupModel?.addLineHighlighter(rowNumber, rowNumber + 1, textAttributes)
-                                    }
-                                } else {
-                                    scroll(0, 0)
-                                }
+                                scroll(0, 0)
                             }
                         }
                     }